the README.md file

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -6,7 +6,7 @@
  */
 	int print_last_digit(int n)
 {
-	long a, b;
+	unsigned int a, b;
 
 	if (n < 0)
 	{
